Modified the activities Added new files.

+++ app/src/main/java/com/example/chsan_000/database/MyShowHistory.java
@@ -1,20 +1,13 @@
 package com.example.chsan_000.database;
 
-import android.content.Context;
-import android.database.DataSetObservable;
 import android.database.sqlite.SQLiteDatabase;
 import android.os.Bundle;
 import android.support.v7.app.ActionBar;
 import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.DefaultItemAnimator;
 import android.support.v7.widget.LinearLayoutManager;
 import android.support.v7.widget.RecyclerView;
-import android.support.v7.widget.StaggeredGridLayoutManager;
-import android.util.Log;
 import android.view.Menu;
 import android.view.MenuItem;
-import android.view.View;
-import android.widget.Button;
 import android.widget.ProgressBar;
 import android.widget.Toast;
 
@@ -28,77 +21,63 @@ import java.util.List;
  */
 public class MyShowHistory extends AppCompatActivity {
     private MyHistoryAdapter empAdap;
-    private ProgressBar progressBar;
+    ProgressBar progressBar;
     private RecyclerView recyclerView;
-    private Button button;
-    private List<PostPojo> employeeList = new ArrayList<>();
+    private ActionBar actionBar;
+    List<PostPojo> employeeList = new ArrayList<>();
 
-    final String TAG = "DatabaseViewRecordsActivity";
+    private static final String TAG = "MyShowHistory";
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_recycle);
 
-        ActionBar actionBar = getSupportActionBar();
-        //actionBar.setDisplayHomeAsUpEnabled(true);
+        actionBar = getSupportActionBar();
         actionBar.setDisplayShowHomeEnabled(true);
         //actionBar.setIcon(R.drawable.cloud);
         actionBar.setTitle("History");
 
         recyclerView = (RecyclerView) findViewById(R.id.recycler_view);
         progressBar = (ProgressBar) findViewById(R.id.myProgressBar);
-        button = (Button) findViewById(R.id.my_btn);
-
-        button.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                DbOpenHelper dbOpenHelper = new DbOpenHelper(MyShowHistory.this);
-                SQLiteDatabase db = dbOpenHelper.getWritableDatabase();
 
 
-                db.execSQL("delete from " + dbOpenHelper.TABLE_NAME);
+        recyclerView.setLayoutManager(new LinearLayoutManager(MyShowHistory.this));
+        recyclerView.setHasFixedSize(true);
 
-                db.close();
-
-                empAdap.notifyDataSetChanged();
-
-                Toast.makeText(MyShowHistory.this, "Deleted Successfully", Toast.LENGTH_SHORT).show();
-            }
-        });
-
-        empAdap = new MyHistoryAdapter(employeeList);
-
+        empAdap = new MyHistoryAdapter(employeeList, R.layout.activity_recycle, MyShowHistory.this);
+        empAdap.notifyDataSetChanged();
         recyclerView.setAdapter(empAdap);
-
-        recyclerView.setLayoutManager(new StaggeredGridLayoutManager(2, StaggeredGridLayoutManager.VERTICAL));
-
-        RecyclerView.ItemAnimator itemAnimator = new DefaultItemAnimator();
-        itemAnimator.setAddDuration(1000);
-        itemAnimator.setRemoveDuration(1000);
-        recyclerView.setItemAnimator(itemAnimator);
-
 
         ZipCodeAsync databaseAsync = new ZipCodeAsync(this, progressBar, empAdap);
         databaseAsync.execute("Employee Records");
+        empAdap.addZip(employeeList);
+
+        empAdap.notifyDataSetChanged();
     }
 
-//    @Override
-//    public boolean onCreateOptionsMenu(Menu menu) {
-//        getMenuInflater().inflate(R.menu.menu_history_delete, menu);
-//        return super.onCreateOptionsMenu(menu);
-//    }
-//
-//    @Override
-//    public boolean onOptionsItemSelected(MenuItem item) {
-//        if (item.getItemId() == R.id.delete_history) {
-//            DbOpenHelper dbOpenHelper = new DbOpenHelper(this);
-//            SQLiteDatabase db = dbOpenHelper.getWritableDatabase();
-//            db.execSQL("delete from " + dbOpenHelper.TABLE_NAME);
-//
-//            Toast.makeText(MyShowHistory.this, "Deleted Successfully", Toast.LENGTH_SHORT).show();
-//
-//        }
-//        return super.onOptionsItemSelected(item);
-//    }
+    @Override
+    public boolean onCreateOptionsMenu(Menu menu) {
+        getMenuInflater().inflate(R.menu.menu_history_delete, menu);
+        return super.onCreateOptionsMenu(menu);
+    }
+
+    @Override
+    public boolean onOptionsItemSelected(MenuItem item) {
+        if (item.getItemId() == R.id.delete_history) {
+
+            DbOpenHelper dbOpenHelper = new DbOpenHelper(this);
+            SQLiteDatabase db = dbOpenHelper.getWritableDatabase();
+            db.execSQL("delete from " + dbOpenHelper.TABLE_NAME);
+
+            empAdap.notifyDataSetChanged();
+
+            empAdap = new MyHistoryAdapter(employeeList, R.layout.activity_recycle, MyShowHistory.this);
+            empAdap.notifyDataSetChanged();
+            recyclerView.setAdapter(empAdap);
+
+            Toast.makeText(MyShowHistory.this, "Deleted Successfully", Toast.LENGTH_SHORT).show();
+        }
+        return super.onOptionsItemSelected(item);
+    }
 }
